pyproject.toml
[tool.poetry]
name = "botfaqs"
version = "0.1.0"
description = "A Discord bot for managing and answering frequently asked questions"
authors = ["Rick Anderson <rick.rickanderson@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.1.0"
tortoise-orm = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run = "python src/bot.py"


